requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Data Processing
opencv-python>=4.8.0
albumentations>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# API & UI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
python-multipart>=0.0.6
Pillow>=10.0.0

# Optimization
onnx>=1.14.0
onnxruntime>=1.16.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
scikit-learn>=1.3.0